leaderboard-bot_v2/requirements.txt
kubernetes
python-dotenv
slackclient
psycopg2
boto3 == 1.29